feat: Enhance User and Repository models with discussions, social links, and additional properties (#75)
feat: Add social links generation for repositories in seed data (#78)

Feat/seed likes (#79)

* feat: Integrate RepositoryLeaderboard into DemoRepositoryPage and manage leaderboard state

* feat: Add user favorites creation logic in seed script

refactor: Reorganize imports and improve discussion layout for better readability (#80)

feat: Implement action to retrieve top voted repositories for the current week (#82)

Fix/repo page (#81)

* feat: Integrate RepositoryLeaderboard into DemoRepositoryPage and manage leaderboard state

* refactor: remove demo repository page and layout files; implement new repository page with tabs for leaderboard, discussions, and social links

Fix/repo page (#85)

* feat: Integrate RepositoryLeaderboard into DemoRepositoryPage and manage leaderboard state

* refactor: remove demo repository page and layout files; implement new repository page with tabs for leaderboard, discussions, and social links

* feat: Implement action to retrieve top voted repositories for the current week (#82)

---------


Feat/api shema (#86)

* feat: Integrate RepositoryLeaderboard into DemoRepositoryPage and manage leaderboard state

* feat: add ApiUser and ApiKey models with status enum for API key management

feat: Improve RepoCard styling with hover effects and transparent badge variant (#87)

refactor: remove demo repository page and layout files; implement new repository page with tabs for leaderboard, discussions, and social links
