+++ src/actions/repository/getTopReposThisWeek/index.ts
@@ -1,0 +1,5 @@
+export { getTopReposThisWeekAction } from './action';
+export { getTopReposThisWeek } from './logic';
+export { getTopReposThisWeekSchema } from './schema';
+export type { GetTopReposThisWeekInput } from './schema';
+export type { TopRepoThisWeek, GetTopReposThisWeekResult } from './logic';
